+++ ampd/src/prometheus_metrics/mod.rs
@@ -1,3 +1,2 @@
-mod metrics;
+pub mod metrics;
 pub mod monitor;
-pub mod msg;
